rename demo & docs & fix bugs
