+++ .eslintignore
@@ -1,2 +1,2 @@
 tests
-demo
+site
