+++ package.json
@@ -4,7 +4,7 @@
   "description": "Element UI for React",
   "main": "index.js",
   "scripts": {
-    "start": "cd demo && node run.js",
+    "start": "cd site && node run.js",
     "test": "jest --config jest.json",
     "lint": "eslint -c .eslintrc.json src",
     "lint:src:q": "eslint --quiet -c .eslintrc.json src",
@@ -29,9 +29,9 @@
     "babel-loader": "^6.2.5",
     "babel-preset-es2015": "^6.16.0",
     "babel-preset-react": "^6.16.0",
-    "babel-standalone": "^6.18.0",
+    "babel-standalone": "^6.18.1",
     "css-loader": "^0.25.0",
-    "eslint": "^3.8.1",
+    "eslint": "^3.9.0",
     "eslint-config-airbnb": "^12.0.0",
     "eslint-plugin-import": "^1.16.0",
     "eslint-plugin-jsx-a11y": "^2.2.3",
