+++ site/pages/badge/index.js
